+++ internal/cli/onboard/onboard.go
@@ -7,8 +7,8 @@ import (
 	"github.com/apex/log"
 	"github.com/fatih/color"
 	ooni "github.com/ooni/probe-cli"
-	"github.com/ooni/probe-cli/config"
 	"github.com/ooni/probe-cli/internal/cli/root"
+	"github.com/ooni/probe-cli/internal/config"
 	"github.com/ooni/probe-cli/internal/output"
 	"github.com/pkg/errors"
 	"gopkg.in/AlecAivazis/survey.v1"
