+++ internal/config/testdata/valid-config.json
